+++ go/libraries/doltcore/diff/table_diff_test.go
@@ -1,0 +1,1 @@
+package diff
